implement links

+++ src/components/contents/StoryDetails/TabsContent/Files/index.tsx
@@ -6,6 +6,9 @@ import { api } from '@root/api'
 import { FileType } from '@root/types'
 import { Divider, Header as SHeader } from 'semantic-ui-react'
 import FilesList from './FilesList'
+import LoadingArea from '@root/components/suport/LoadingArea'
+import { useSetRecoilState } from 'recoil'
+import { basicModalState } from '@root/components/modals/BasicModal'
 
 interface iFiles {
   files: FileType[]
@@ -15,39 +18,60 @@ interface iFiles {
 }
 
 const Files: React.FC<iFiles> = ({ storyId, refresh, files, isLoading, children }) => {
-  const handleUploadFinished = async (value: FileType) => {
-    await api.addFileToStory(storyId, value.id)
-    refresh()
-  }
+  const setBasicModalState = useSetRecoilState(basicModalState)
 
   const handleUploadProgress = (value: number) => {
     console.log(value)
   }
 
-  const handleRemoveFile = async (value: FileType) => {
-    await api.removeFileFromStory(storyId, value.id)
-    await api.deleteFile(value.id)
+  const onRemove = async (value: FileType) => {
+    setBasicModalState({
+      open: true,
+      title: 'Remove file',
+      isConfirmation: true,
+      onClose: async (isSuccess: boolean) => {
+        if (!isSuccess) return
+        await api.deleteFile(value.id)
+        refresh()
+      },
+      content: (
+        <>
+          Do you want remove file <b>{value.originalName}</b> from this story?
+        </>
+      )
+    })
+  }
+
+  const handleSelected = async (event: React.ChangeEvent<HTMLInputElement>) => {
+    if (!event.target.files?.length) return
+    const file: File = event.target.files[0]
+    await api.uploadFile(storyId, file, ({ loaded, total }) => {
+      const progress = Math.round((loaded * 100) / total) - 1
+      console.log(progress < 0 ? 0 : progress)
+    })
     refresh()
   }
 
   return (
     <Layout>
-      <ColumnForm>
-        <Header>Link files to your video for viewers to download</Header>
-        <UploadButton
-          api={api.uploadFile}
-          onFinished={handleUploadFinished}
-          onProgressChange={handleUploadProgress}
-          basic
-          primary
-        >
-          Upload new file
-        </UploadButton>
-        <Divider />
-        <SHeader as="h1">List of files</SHeader>
-        <FilesList files={files} onRemove={handleRemoveFile} />
-      </ColumnForm>
-      <ColumnPreview>{children}</ColumnPreview>
+      <LoadingArea isLoading={isLoading}>
+        <ColumnForm>
+          <Header>Link files to your video for viewers to download</Header>
+          <UploadButton
+            api={api.uploadFile}
+            onSelected={handleSelected}
+            onProgressChange={handleUploadProgress}
+            basic
+            primary
+          >
+            Upload new file
+          </UploadButton>
+          <Divider />
+          <SHeader as="h1">List of files</SHeader>
+          <FilesList files={files} onRemove={onRemove} />
+        </ColumnForm>
+        <ColumnPreview>{children}</ColumnPreview>
+      </LoadingArea>
     </Layout>
   )
 }
